pyproject.toml
[project]
name = "stylegan3"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "Takeru Saito", email = "takelushi@gmail.com" }]
dependencies = [
    "numpy>=1.20",
    "click>=8.1",
    "pillow>=10.2.0",
    "scipy>=1.10.1",
    "requests>=2.28.1",
    "tqdm>=4.64.1",
    "ninja==1.10.2",
    "matplotlib==3.4.2",
    "imageio==2.9.0",
    "imgui==1.3.0",
    "glfw==2.2.0",
    "pyopengl==3.1.5",
    "imageio-ffmpeg==0.4.3",
    "pyspng>=0.1.1",
    "setuptools==69.5.1",
    "torch==2.3.1+cu121",
    "psutil>=6.0.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[[tool.rye.sources]]
name = "torch"
url = "https://download.pytorch.org/whl/cu121"
type = "index"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/stylegan3"]
